Rename Namespace and file
